refactor: DailyNutrition의 DataModel을 파일별로 나눴습니다. (SCRUM-117)
